requirements.txt
youtube_transcript_api
streamlit
streamlit-lottie
google-generativeai
python-dotenv
pathlib
